+++ gemini-android/src/main/java/com/haroncode/gemini/android/AndroidBinder.kt
@@ -1,13 +1,15 @@
-package com.haroncode.gemini.android.binder
+package com.haroncode.gemini.android
 
 import androidx.lifecycle.LifecycleOwner
-import com.haroncode.gemini.android.binder.strategies.StartStopStrategy
+import com.haroncode.gemini.android.binder.MainThreadConnectionBinder
+import com.haroncode.gemini.android.lifecycle.AndroidStoreLifecycle
+import com.haroncode.gemini.android.strategies.StartStopStrategy
 import com.haroncode.gemini.binder.BaseConnectionBinder
 import io.reactivex.disposables.Disposable
 
 object AndroidBinder {
 
-    fun <T : LifecycleOwner> withBinding(viewBinding: ViewBinding<T>): Builder<T> = Builder(viewBinding)
+    fun <T : LifecycleOwner> withBinding(viewBinding: ViewBinding<T>) = Builder(viewBinding)
 
     class Builder<T : LifecycleOwner>(private val viewBinding: ViewBinding<T>) {
 
@@ -19,10 +21,14 @@ object AndroidBinder {
         }
 
         fun create(view: T): Disposable {
-            val lifecycle = AndroidStoreLifecycle(view.lifecycle, bindingStrategy)
+            val lifecycle = AndroidStoreLifecycle(
+                view.lifecycle,
+                bindingStrategy
+            )
 
             val baseConnectionBinder = BaseConnectionBinder(lifecycle)
-            val mainThreadConnectionBinder = MainThreadConnectionBinder(baseConnectionBinder)
+            val mainThreadConnectionBinder =
+                MainThreadConnectionBinder(baseConnectionBinder)
 
             viewBinding.onCreate(view, mainThreadConnectionBinder)
             return mainThreadConnectionBinder
